requirements.txt
streamlit
pandas
numpy
scikit-learn
requests
plotly
matplotlib
seaborn
google-generativeai
